Moving Git::Base factory methods into a module
